+++ cmd/neofs-lens/internal/storage/root.go
@@ -36,13 +36,14 @@ var (
 
 var Root = &cobra.Command{
 	Use:   "storage",
-	Short: "Operations with an object",
+	Short: "Operations with blobstor",
 	Args:  cobra.NoArgs,
 }
 
 func init() {
 	Root.AddCommand(
 		storageGetObjCMD,
+		storageListObjsCMD,
 		storageStatusObjCMD,
 	)
 }